docs/context/copilot_agent_starter/copilot_agent/pyproject.toml
[project]
name = "copilot-agent"
version = "0.1.0"
description = "Minimal Copilot Agent (FastAPI + tools: RAG, reports)"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
    "duckdb>=1.1.2",
    "pandas>=2.2.2",
    "numpy>=2.0.2",
    "faiss-cpu>=1.8.0.post1",
    "sentence-transformers>=3.0.1",
    "tiktoken>=0.7.0"
]

[project.optional-dependencies]
openai = ["openai>=1.40.0"]

[tool.uv]
# ensures uv respects project layout

[tool.black]
line-length = 100
